Keep all the values, for enhanced analysis.
Doing running min/max/total is easy, but we'll want to add more complex
analysis so keep all the values.

This also means we keep the original text, which makes it much neater
when we want to dump out numbers which didn't change, and also to tell if
we should insert a space before numbers.

Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>